requirements.txt
tensorflow>=2.15.0
numpy
pillow
opencv-python
scikit-learn
matplotlib
streamlit
datasets